project completed

+++ backend/controllers/product.controller.js
@@ -53,11 +53,15 @@ export const createProducts = async(req,res)=>{
 export const deleteProducts = async(req,res)=>{
     const {id} = req.params
     
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({success:false,massage:"Invalid Product Id"});
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success:true , message:"Product deleted"});
     } catch (error) {
         console.log("Error in deleting product:",error.massage);
-        res.status(404).json({success:false,massage:"Product not found"});
+        res.status(500).json({success:false,massage:"Server Error"});
     }
 };
